backend/src/utils/hashing.py
from argon2 import PasswordHasher
from argon2.exceptions import VerifyMismatchError, VerificationError, InvalidHashError

ph = PasswordHasher(
    time_cost=3,
    memory_cost=65536,
    parallelism=2,
)


def hash_password(plain: str) -> str:
    return ph.hash(plain)


def verify_password(plain: str, hashed: str) -> bool:
    try:
        return ph.verify(hashed, plain)
    except (VerifyMismatchError, VerificationError, InvalidHashError):
        return False


def hash_token(token: str) -> str:
    return ph.hash(token)


def verify_token(token: str, hashed: str) -> bool:
    try:
        return ph.verify(hashed, token)
    except (VerifyMismatchError, VerificationError, InvalidHashError):
        return False