backend/pyproject.toml
[project]
name = "car-rental-backend"
version = "0.1.0"
description = "A FastAPI backend for a car rental system with LangGraph integration."
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "asyncpg>=0.29.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "aiobotocore[boto3]>=2.7.0",
    "celery[redis]>=5.3.0",
    "redis>=5.0.0",
    "argon2-cffi>=23.1.0",
    "langgraph>=0.1.0",
    "langchain>=0.1.0",
    "slowapi>=0.1.9",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
    "Pillow>=10.0.0",
]

[dependency-groups]
dev = ["pytest", "pytest-asyncio", "ruff", "mypy"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]